Distributed bench (#1661)
* Remote benchmarking

+++ sui/src/benchmark/validator_preparer.rs
@@ -2,12 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #![allow(clippy::large_enum_variant)]
+use crate::benchmark::bench_types::RunningMode;
+use crate::benchmark::load_generator::spawn_authority_server;
+use crate::config::NetworkConfig;
+use crate::config::{AccountConfig, Config, GenesisConfig, ObjectConfig};
 
-use crate::{
-    benchmark::{bench_types::RunningMode, load_generator::spawn_authority_server},
-    config::{AccountConfig, Config, GenesisConfig, ObjectConfig},
-};
 use rocksdb::Options;
+use std::collections::BTreeMap;
 use std::{
     env, fs, panic,
     path::{Path, PathBuf},
@@ -77,10 +78,26 @@ pub enum ValidatorConfig {
         working_dir: PathBuf,
         validator_process: Option<Child>,
     },
+    RemoteValidatorConfig,
 }
 
 impl ValidatorPreparer {
-    pub fn new(
+    pub fn new_for_remote(
+        network_config: &NetworkConfig,
+        keys: &BTreeMap<PublicKeyBytes, KeyPair>,
+    ) -> Self {
+        let keys = keys.iter().map(|q| q.1.copy()).collect::<Vec<_>>();
+        let committee = Committee::from(network_config);
+
+        Self {
+            running_mode: RunningMode::RemoteValidator,
+            keys,
+            main_authority_address_hex: "".to_string(),
+            committee,
+            validator_config: ValidatorConfig::RemoteValidatorConfig,
+        }
+    }
+    pub fn new_for_local(
         running_mode: RunningMode,
         working_dir: Option<PathBuf>,
         committee_size: usize,
@@ -107,7 +124,7 @@ impl ValidatorPreparer {
         );
 
         match running_mode {
-            RunningMode::LocalSingleValidatorProcess => {
+            RunningMode::SingleValidatorProcess => {
                 // Honor benchmark's host:port setting
                 genesis_config.authorities[0].port = validator_port;
                 genesis_config.authorities[0].host = validator_host.into();
@@ -124,7 +141,7 @@ impl ValidatorPreparer {
                 }
             }
 
-            RunningMode::LocalSingleValidatorThread => {
+            RunningMode::SingleValidatorThread => {
                 // Pick the first validator and create state.
                 let public_auth0 = keys[0].public_key_bytes();
                 let secret_auth0 = keys[0].copy();
@@ -150,12 +167,13 @@ impl ValidatorPreparer {
                     },
                 }
             }
+            RunningMode::RemoteValidator => panic!("Use new_for_remote"),
         }
     }
 
     pub fn deploy_validator(&mut self, _network_server: NetworkServer) {
         match self.running_mode {
-            RunningMode::LocalSingleValidatorProcess => {
+            RunningMode::SingleValidatorProcess => {
                 if let ValidatorConfig::LocalSingleValidatorProcessConfig {
                     working_dir,
                     genesis_config,
@@ -182,7 +200,7 @@ impl ValidatorPreparer {
                     panic!("Invalid validator config in local-single-validator-process mode");
                 }
             }
-            RunningMode::LocalSingleValidatorThread => {
+            RunningMode::SingleValidatorThread => {
                 if let ValidatorConfig::LocalSingleValidatorThreadConfig {
                     authority_state,
                     authority_store: _,
@@ -203,6 +221,7 @@ impl ValidatorPreparer {
                     panic!("Invalid validator config in local-single-validator-thread mode");
                 }
             }
+            RunningMode::RemoteValidator => (),
         }
         // Wait for server start
         sleep(Duration::from_secs(3));
@@ -210,7 +229,7 @@ impl ValidatorPreparer {
 
     pub fn update_objects_for_validator(&mut self, objects: Vec<Object>, address: SuiAddress) {
         match self.running_mode {
-            RunningMode::LocalSingleValidatorProcess => {
+            RunningMode::SingleValidatorProcess => {
                 let all_objects: Vec<ObjectConfig> = objects
                     .iter()
                     .map(|object| ObjectConfig {
@@ -231,12 +250,13 @@ impl ValidatorPreparer {
                         .push(AccountConfig {
                             address: Some(address),
                             gas_objects: all_objects,
+                            gas_object_ranges: None,
                         })
                 } else {
                     panic!("invalid validator config in local-single-validator-process mode");
                 }
             }
-            RunningMode::LocalSingleValidatorThread => {
+            RunningMode::SingleValidatorThread => {
                 if let ValidatorConfig::LocalSingleValidatorThreadConfig {
                     authority_state: _,
                     authority_store,
@@ -249,29 +269,29 @@ impl ValidatorPreparer {
                     panic!("invalid validator config in local-single-validator-thread mode");
                 }
             }
+
+            // Nothing to do here. Remote machine must be provisioned separately
+            RunningMode::RemoteValidator => (),
         }
     }
 
     pub fn clean_up(&mut self) {
-        match self.running_mode {
-            RunningMode::LocalSingleValidatorProcess => {
-                info!("Cleaning up local validator process...");
-                if let ValidatorConfig::LocalSingleValidatorProcessConfig {
-                    working_dir: _,
-                    genesis_config: _,
-                    validator_process,
-                } = &mut self.validator_config
-                {
-                    validator_process
-                        .take()
-                        .unwrap()
-                        .kill()
-                        .expect("Failed to kill validator process");
-                } else {
-                    panic!("invalid validator config in local-single-validator-process mode");
-                }
+        if let RunningMode::SingleValidatorProcess = self.running_mode {
+            info!("Cleaning up local validator process...");
+            if let ValidatorConfig::LocalSingleValidatorProcessConfig {
+                working_dir: _,
+                genesis_config: _,
+                validator_process,
+            } = &mut self.validator_config
+            {
+                validator_process
+                    .take()
+                    .unwrap()
+                    .kill()
+                    .expect("Failed to kill validator process");
+            } else {
+                panic!("invalid validator config in local-single-validator-process mode");
             }
-            RunningMode::LocalSingleValidatorThread => {}
         }
     }
 }
